Attempt at using local storage: didn't work
Tutorial is broken. This is my attempt at it by following directions.
Also I did copy-paste the full source code at the end of the tutorial
and it doesn't work either: it saves huge amount of stuff and also todos
appear blank on the view: it seems the structure of what's saved is not
the one the app expects.